% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{colormap_pal}
\alias{colormap_pal}
\title{Create a Palette generating function}
\usage{
colormap_pal(alpha = 1, colormap = colormaps$jet, reverse = FALSE)
}
\arguments{
\item{alpha}{pass through parameter to \code{colormap}}

\item{colormap}{pass through parameter to \code{colormap}}

\item{reverse}{pass through parameter to \code{colormap}}
}
\value{
A function that can generate colors from a specified colormap.
}
\description{
Create a Palette generating function
}
\examples{
scales::show_col(colormap_pal()(10))
scales::show_col(colormap_pal(colormap=colormaps$viridis)(100), labels=FALSE)

}

