% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colormap.R
\docType{data}
\name{colormaps}
\alias{colormaps}
\title{List of pre-defined colormaps}
\format{An object of class \code{list} of length 44.}
\usage{
colormaps
}
\description{
List of pre-defined colormaps
}
\keyword{datasets}

