% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_colormap}
\alias{scale_color_colormap}
\alias{scale_colour_colormap}
\alias{scale_fill_colormap}
\title{Colormap color scales}
\usage{
scale_color_colormap(..., alpha = 1, colormap = colormaps$jet,
  discrete = FALSE, reverse = FALSE)

scale_fill_colormap(..., alpha = 1, colormap = colormaps$jet,
  discrete = FALSE, reverse = FALSE)
}
\arguments{
\item{...}{parameters to \code{discrete_scale} or \code{scale_fill_gradientn}}

\item{alpha}{pass through parameter to \code{colormap}}

\item{colormap}{pass through parameter to \code{colormap}}

\item{discrete}{generate a discrete palette? (default: \code{FALSE} - generate continuous palette)}

\item{reverse}{pass through parameter to \code{colormap}}
}
\description{
Uses the colormap color scale
}
\details{
For \code{discrete == FALSE} (the default) all other arguments are as to
\link[ggplot2]{scale_fill_gradientn} or \link[ggplot2]{scale_color_gradientn}.
Otherwise the function will return a \code{discrete_scale} with the plot-computed
number of colors.

See \link[colormap]{colormap} for more information on the color scale.
}

